refactor: 💡 code reorganization
