+++ packages/cli/scripts/lib/cliUtils.js
@@ -1,5 +1,6 @@
 const path = require('path');
-const fs = require('fs');
+
+const webpackConfig = require('../../webpack/config');
 
 function statsFormattedOutput(err, stats) {
   if (!err) {
@@ -52,57 +53,21 @@ function builderFactory(cliOptions = {}) {
   };
 }
 
-function requireConfig({
-  rootDir,
-  fileNames,
-  packageJson = null,
-  packageJsonKey = '',
-  defaultConfig = {}
-}) {
-  if (
-    !rootDir ||
-    !Array.isArray(fileNames) ||
-    fileNames.length === 0 ||
-    !fs.existsSync(rootDir)
-  ) {
-    return defaultConfig;
-  }
-
-  const { fullPath: configPath, fileName: configFileName } =
-    fileNames
-      .map(fileName => ({
-        fileName,
-        fullPath: path.resolve(rootDir, fileName)
-      }))
-      .find(({ fullPath }) => fs.existsSync(fullPath)) || {};
-
-  if (
-    !configPath &&
-    !(packageJson && packageJsonKey && packageJson[packageJsonKey])
-  ) {
-    return defaultConfig;
-  }
-
-  try {
-    if (configPath) {
-      const extension = configFileName.split('.').pop();
-
-      return ~['js', 'cjs', 'json'].indexOf(extension)
-        ? require(configPath)
-        : JSON.parse(fs.readFileSync(configPath));
-    } else {
-      return packageJson[packageJsonKey];
-    }
-  } catch (error) {
-    console.error(`Error occurred while loading ${configPath} file.`);
-    console.error(error);
-
-    return defaultConfig;
-  }
+async function createWebpackConfig(args) {
+  return [
+    await webpackConfig({
+      ...args,
+      isServer: true
+    }),
+    await webpackConfig({
+      ...args,
+      isServer: false
+    })
+  ];
 }
 
 module.exports = {
-  requireConfig,
+  createWebpackConfig,
   statsFormattedOutput,
   handlerFactory,
   builderFactory
