+++ packages/cli/scripts/start.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const childProcess = require('child_process');
 
-const { handlerFactory, builderFactory } = require('../build/utils');
+const { handlerFactory, builderFactory } = require('./lib/cliUtils');
 
 async function start({ rootDir }) {
   childProcess.fork(path.resolve(rootDir, './build/server'));
